safari-534.53-branch/LayoutTests/svg/dynamic-updates/SVGFESpotLightElement-svgdom-y-prop-expected.txt

CONSOLE MESSAGE: line 44: TypeError: 'undefined' is not an object (evaluating 'spotLightElement.y.baseVal = 0')
FAIL: Timed out waiting for notifyDone to be called
SVG 1.1 dynamic update tests

Tests dynamic updates of the 'y' property of the SVGFESpotLightElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL spotLightElement.y.baseVal should be 100. Threw exception TypeError: 'undefined' is not an object (evaluating 'spotLightElement.y.baseVal')
